HW2+optional

+++ src/main/java/ru/javawebinar/topjava/repository/inmemory/InMemoryMealRepositoryImpl.java
@@ -1,14 +1,19 @@
 package ru.javawebinar.topjava.repository.inmemory;
 
+import org.springframework.stereotype.Repository;
 import ru.javawebinar.topjava.model.Meal;
 import ru.javawebinar.topjava.repository.MealRepository;
 import ru.javawebinar.topjava.util.MealsUtil;
+import ru.javawebinar.topjava.util.exception.NotFoundException;
 
 import java.util.Collection;
+import java.util.Comparator;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.stream.Collectors;
 
+@Repository
 public class InMemoryMealRepositoryImpl implements MealRepository {
     private Map<Integer, Meal> repository = new ConcurrentHashMap<>();
     private AtomicInteger counter = new AtomicInteger(0);
@@ -17,8 +22,7 @@ public class InMemoryMealRepositoryImpl implements MealRepository {
         MealsUtil.MEALS.forEach(this::save);
     }
 
-    @Override
-    public Meal save(Meal meal) {
+    private Meal save(Meal meal) {
         if (meal.isNew()) {
             meal.setId(counter.incrementAndGet());
             repository.put(meal.getId(), meal);
@@ -29,18 +33,35 @@ public class InMemoryMealRepositoryImpl implements MealRepository {
     }
 
     @Override
-    public void delete(int id) {
-        repository.remove(id);
+    public Meal save(Meal meal, int userId) {
+        if(meal.getUserId()==0||meal.getUserId()==userId){
+            meal.setUserId(userId);
+            return save(meal);
+        }
+       throw new NotFoundException("Security error");
     }
 
     @Override
-    public Meal get(int id) {
-        return repository.get(id);
+    public void delete(int id, int userId) {
+        if(repository.get(id).getUserId()==userId){repository.remove(id);}
+        else {throw new NotFoundException("Security error delete");}
     }
 
     @Override
-    public Collection<Meal> getAll() {
-        return repository.values();
+    public Meal get(int id, int userId) {
+        if(repository.get(id).getUserId()==userId){
+        return repository.get(id);}
+       else {
+           return null;
+        }
+    }
+
+    @Override
+    public Collection<Meal> getAll(int userId) {
+        return repository.values().stream()
+                .filter(x->x.getUserId()==userId)
+                .sorted(Comparator.comparing(Meal::getDateTime))
+                .collect(Collectors.toList());
     }
 }
